feat(alchm): restore legacy home order, move dashboard to /lab
The Phase 2 dashboard replaced the original home page entirely, losing
the promo + recommender stack that had organized this site for over a
year. This commit conserves both:

Home page (/)
- Stacked single-column layout in the alchm style, mirroring the
  long-established legacy order:
  1. AmazonFreshPromotion (Cosmic Token Economy)
  2. Cuisine + Sauce: DynamicCuisineRecommender
  3. Ingredient recommender (lite): top-8 IngredientCards backed by
     /api/recommendations/ingredients, with a "Full recommender" CTA
     linking to /ingredients
  4. Cooking methods: EnhancedCookingMethodRecommender
- Footer chip points at /lab for the laboratory view.

/lab
- The Phase 2 dashboard (PlanetaryClock hero, natal rail, cuisine
  explorer, sauce lineage, telemetry) moves here verbatim. Nothing is
  lost — it is just no longer the front door.

LabHeader
- Nav expanded from 5 to 9 entries: Kitchen, Ingredients, Cuisines,
  Methods, Pantry, Meal Plan, Recipes, Profile, Lab. Each links to its
  existing route.
- Now horizontally scrollable on mobile so the full set is reachable
  at any width; centers in a pill on ≥1100px.
- Active-route detection reads usePathname() so every sub-route lights
  up its parent tab automatically; the hard-coded `active="kitchen"`
  in (alchm)/layout.tsx is removed.
